Refactor button so that it can be used in more places
