+++ apps/docs/src/routes/_docs/komponenter/$slug.tsx
@@ -3,11 +3,11 @@ import { AnchorHeading } from '@/ui/anchor-heading';
 import { PropsTable } from '@/ui/props-table';
 import { ResourceLink, ResourceLinks } from '@/ui/resource-links';
 import { SanityContent } from '@/ui/sanity-content';
+import { ScrollToTop } from '@/ui/scroll-to-top';
 import { TableOfContentsNav } from '@/ui/table-of-contents-nav';
-import { ArrowUp, Child, CircusTent } from '@obosbbl/grunnmuren-icons-react';
+import { Child, CircusTent } from '@obosbbl/grunnmuren-icons-react';
 import { Alertbox, Content } from '@obosbbl/grunnmuren-react';
 import { createFileRoute, notFound } from '@tanstack/react-router';
-import { useEffect, useState } from 'react';
 import type * as props from 'component-props';
 import { defineQuery } from 'groq';
 
@@ -46,25 +46,6 @@ export const Route = createFileRoute('/_docs/komponenter/$slug')({
 
 function Page() {
   const { data } = Route.useLoaderData();
-  const [showScrollToTop, setShowScrollToTop] = useState(false);
-
-  // Show/hide scroll to top button based on scroll position
-  useEffect(() => {
-    const handleScroll = () => {
-      const scrollTop = window.scrollY || document.documentElement.scrollTop;
-      setShowScrollToTop(scrollTop > 300); // Show after scrolling 300px
-    };
-
-    window.addEventListener('scroll', handleScroll);
-    return () => window.removeEventListener('scroll', handleScroll);
-  }, []);
-
-  const scrollToTop = () => {
-    window.scrollTo({
-      top: 0,
-      behavior: 'smooth',
-    });
-  };
 
   const ghLink = data.resourceLinks?.find(
     (link) => link.linkType === 'github',
@@ -157,20 +138,7 @@ function Page() {
         </div>
       </div>
 
-      {/* Scroll to top button */}
-      {showScrollToTop && (
-        <div className="fixed right-4 bottom-4 z-50 flex flex-col items-center md:right-16 md:bottom-16">
-          <button
-            onClick={scrollToTop}
-            className="hover:-translate-y-1 flex h-12 w-12 items-center justify-center rounded-full bg-blue-dark transition-all duration-300 focus-visible:outline-focus"
-            aria-label="Scroll to top"
-            type="button"
-          >
-            <ArrowUp className="h-6 w-6 text-white" />
-          </button>
-          <span className="mt-2 font-medium">Til tops</span>
-        </div>
-      )}
+      <ScrollToTop />
     </>
   );
 }
